Use different path for temporary files
The default path (basically "/tmp") is now used for temporary files used
during the image creation process, because by using "/data" inside the
Docker container, we were relying on the filesystem outside of the
container too much.

For example, when using WSL and accessing the host's file system, the
issue was that deleted files could not be reopened via "/dev/fd/...",
while everything would work fine when directly working on WSL's ext4
filesystem.

This resolves issue #2.